Squashed 'external/PlayfieldEngine/' changes from f5dd818..035f5a3
035f5a3 Fix a conversion warning
d588151 Fix Clang warnings
2e817d8 Improve sysbtn behavior
5a46598 Merge upstream
f580d8e Merge upstream
ed04187 Merge upstream
2cfae1f Merge upstream
03b1f58 Merge upstream
0e1fe26 Merge upstream
e55d436 Merge upstream
9bc265a Merge upstream
46751f7 Fix save UI
fd295b1 Merge upstream
6e94bdc Megre upstream
ae3a3e0 Merge upstream
0a35be9 Merge upstream
d0e6217 Merge upstream
77fe0b4 Merge upstream
744bc9a Merge upstream
7728c36 Merge upstream
37ce19c Merge upstream
755e1f1 Merge upstream
b2d69fe Merge upstream
339ba41 Fix CI
99ab570 Merge upstream
1f442de Resolve conflicts
6fbe10c Merge upstream
da60a24 Merge upstream
f3613d2 Merge upstream
7da8630 Merge upstream
cdb5d1f Merge upstream
c027442 Merge upstream
6e96360 Merge upstream
023cc66 Merge upstream
1b5a65d Merge upstream
9de1bed Fix CI
0ee5e40 Merge upstream
80ae52f Merge upstream
14fee0d Merge upstream
d7780d8 Merge upstream
b7c541e Merge upstream
353ac24 Merge upstream
eb7039d Merge upstream
0532ea9 Merge conflicts
4ecad7b Merge upstream
2c07e9e Merge upstream
3edaae7 Merge upstream
5638afc Merge commit 'fa7d21a56f269862d57a2d28568144099c820db6' as 'external/PlayfieldEngine'
720d2ec wip
d5652de Update README.md
5cd85e6 Update for OpenHarmony
291ebdb Update for OpenHarmony
d864646 Update for OpenHarmony
3b781da Update for OpenHarmony
631a454 Update README.md
903a116 Update README.md
2eb0e74 Update README.md
afe6723 Update README.md
88f278f Update README.md (Add Discord link)
64eb1be Update README.md (Add Discord link)
042b1f8 Update README.md
2d93e10 Update README.md
ee7a968 Update README.md
6190078 Update README.md
8051515 Update README.md
9a50dd4 Update README.md
5f07471 Add logo-small.png
182e7ab Update README.md
d3fe204 Update README.md
df169f1 Update api.c
e94cb36 Fix API consistency
c6596ee Updated api.md
b61eb1b Update api.c
b6caa68 Update api.md and api.c
5cbfdbb Update api.md
3a8b5cb Update README.md
5948d26 Update api.c
21b8a02 Update api.md
3a35127 Update api.c
a244fbb Update api.c
fcabe0f Update api.md
08f5729 Implement many API wrappers. (wip)
dd6b9be Update README.md
3f43536 Updated the API document
76a2c65 Update README.md
25171db Update the tag document
f5b365c Update README.md
ba9f3db Add intro.md
3210abd Update README.md
2307163 Update documents
a9f147c Update README.md
402304a Initial

git-subtree-dir: external/PlayfieldEngine
git-subtree-split: 035f5a3b789255cf16f4bd08172f0c24271d1898

+++ include/playfield/c89compat.h
@@ -471,8 +471,8 @@ static INLINE uint32_t pf_be_to_host_32(uint32_t d) {
 		((d >> 24) & 0xff);
 }
 static INLINE uint16_t pf_be_to_host_16(uint16_t d) {
-	return ((d & 0xff) << 8) |
-		((d >> 8) & 0xff);
+	return (uint16_t)(((uint32_t)(d & 0xff) << 8) |
+		          ((d >> 8) & 0xff));
 }
 static INLINE uint64_t pf_host_to_be_64(uint64_t d) {
 	return ((d & 0xff) << 56) |
@@ -491,8 +491,8 @@ static INLINE uint32_t pf_host_to_be_32(uint32_t d) {
 		((d >> 24) & 0xff);
 }
 static INLINE uint16_t pf_host_to_be_16(uint16_t d) {
-	return ((d & 0xff) << 8) |
-		((d >> 8) & 0xff);
+	return (uint16_t)(((uint32_t)(d & 0xff) << 8) |
+			  ((d >> 8) & 0xff));
 }
 
 #else
